01_basic_agent/requirements.txt
agno>=1.7.0
openai>=1.0.0
python-dotenv>=1.0.0
